refactor: structure of the sign in page
